Move filter types to own file
