+++ moq-transport/src/message/mod.rs
@@ -35,6 +35,7 @@ mod announce;
 mod announce_cancel;
 mod announce_error;
 mod announce_ok;
+mod filter_type;
 mod go_away;
 mod publisher;
 mod subscribe;
@@ -51,6 +52,7 @@ pub use announce::*;
 pub use announce_cancel::*;
 pub use announce_error::*;
 pub use announce_ok::*;
+pub use filter_type::*;
 pub use go_away::*;
 pub use publisher::*;
 pub use subscribe::*;
